Fix various PEP-8 warnings

+++ giftr/__init__.py
@@ -4,32 +4,25 @@ application factory
 import os
 
 from flask import Flask, request, jsonify
-from flask_sqlalchemy import SQLAlchemy
-from sqlalchemy import exc as sqlalchemy_exc
+
 
 from .models import db
 from .exceptions import SetNotFoundError, BadFormatError, DBError
 from . import db_helper
+
 
 def trace():
     from inspect import currentframe, getframeinfo
     cf = currentframe()
     print('>>>>> TRACE: {}:{}'.format(getframeinfo(cf).filename, cf.f_back.f_lineno))
 
-def create_app(test_config=None):
+
+def create_app():
     """
     create and configure the app
     """
     app = Flask(__name__, instance_relative_config=True)
-    
-    use_postgress = True
-    if use_postgress: 
-        # db path for postgres
-        app.config['SQLALCHEMY_DATABASE_URI'] = "postgresql:///gifts"
-    else:
-        # db path for sqlite
-        db_path  = os.path.join(app.instance_path, 'gifts.sqlite')
-        app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///{}'.format(db_path)
+    app.config.from_envvar('GIFTS_SETTINGS')
     
     app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
     db.init_app(app)
@@ -39,21 +32,21 @@ def create_app(test_config=None):
         os.makedirs(app.instance_path)
     except OSError:
         pass
-    
-    
+
     # test interface  - init db
     @app.route('/test',  methods=['POST'])
     def test():
+        if not app.config['TESTING']:
+            return "Not Found", 404
         print("TEST_INTERFACE ")
         data = request.get_json()
-        action = data.get("action","")
+        action = data.get("action", "")
         if action == 'init':
             db.drop_all()
             db.create_all()
-            return("Initialized the database.")
+            return 'Initialized the database.'
         return 'Nothing has been done'
-    
-        
+
     @app.route('/imports', methods=['POST'])
     def insert():
         """
@@ -62,7 +55,8 @@ def create_app(test_config=None):
         Returns: 
             response: response containing import id,  201: Created -  if insertion was successful
             return_str: error message, 400: Bad Request - if can't make insertion due to some problem with client's data
-            return_str: error message, 500: Internal Server Error - if unexpected error occured during insertion (indicator that something is wrong with server)
+            return_str: error message, 500: Internal Server Error - if unexpected error occurred during insertion
+            (indicator that something is wrong with server)
            
         """
         if request.method == 'POST':
@@ -70,11 +64,11 @@ def create_app(test_config=None):
             try:
                 import_id = db_helper.insert_citizens_set(request_json)
                 response = jsonify({"data": {"import_id": import_id}})
-                return response , 201
+                return response, 201
             except (BadFormatError, DBError) as e:
                 return_str = "Insertion failed: {}".format(str(e))
                 return return_str, 400
-            #non-expected exception
+            # non-expected exception
             except Exception as e:
                 trace()
                 import traceback
@@ -93,7 +87,8 @@ def create_app(test_config=None):
         Returns: 
             response: response containing set of citizens,  200: OK -  if query was successful
             return_str: error message, 404: Not Found - if there are no set of citizens with import_id in db
-            return_str: error message, 500: Internal Server Error - if unexpected error occured during query (indicator that something is wrong with server)
+            return_str: error message, 500: Internal Server Error - if unexpected error occurred during query
+            (indicator that something is wrong with server)
         """
         try:
             res = db_helper.get_citizens_set(import_id)
@@ -112,8 +107,7 @@ def create_app(test_config=None):
             return_str = "Insertion failed: {}".format(str(e))
             return return_str, 500
         
-        
-    @app.route('/imports/<int:import_id>/citizens/<int:citizen_id>',methods=['PATCH'])
+    @app.route('/imports/<int:import_id>/citizens/<int:citizen_id>', methods=['PATCH'])
     def patch(import_id, citizen_id):
         """
         Patch interface
@@ -123,10 +117,12 @@ def create_app(test_config=None):
             citizen_id - citizen to patch
            
         Returns: 
-            response: response containing new information about citizen,  200: OK -  if patch was succesfully performed
-            return_str: error message, 404: Not Found - if there is no set of citizens with import_id in db, or there is not citizen with citizen_id in the set
+            response: response containing new information about citizen,  200: OK -  if patch was successfully performed
+            return_str: error message, 404: Not Found - if there is no set of citizens with import_id in db,
+            or there is not citizen with citizen_id in the set
             return_str: error message, 400: Bad Request - if can't perform patch due to some problem with client's data
-            return_str: error message, 500: Internal Server Error - if unexpected error occured during patch performance (indicator that something is wrong with server)
+            return_str: error message, 500: Internal Server Error - if unexpected error occurred during
+            patch performance (indicator that something is wrong with server)
         """
         if request.method == 'PATCH':
             request_json = request.get_json()
@@ -159,9 +155,11 @@ def create_app(test_config=None):
             import_id - id of citizens' set
         
         Returns: 
-            response: response containing information about quantity of presents citizens buy,  200: OK -  if query was successful
+            response: response containing information about quantity of presents citizens buy,  200: OK
+            - if query was successful
             return_str: error message, 404: Not Found - if there are no set of citizens with import_id in db
-            return_str: error message, 500: Internal Server Error - if unexpected error occured during query (indicator that something is wrong with server)
+            return_str: error message, 500: Internal Server Error - if unexpected error occurred during query
+            (indicator that something is wrong with server)
         """
         try:
             res = db_helper.get_citizens_birthdays_for_import_id(import_id)
@@ -182,7 +180,7 @@ def create_app(test_config=None):
     @app.route('/imports/<import_id>/towns/stat/percentile/age')
     def get_statistic(import_id):
         """
-        Interface to get percetiles
+        Interface to get percentiles
            
         Args:
             import_id - id of citizens' set
@@ -190,7 +188,8 @@ def create_app(test_config=None):
         Returns: 
             response: response containing structure with percentiles,  200: OK -  if query was successful
             return_str: error message, 404: Not Found - if there are no set of citizens with import_id in db
-            return_str: error message, 500: Internal Server Error - if unexpected error occured during query (indicator that something is wrong with server code)
+            return_str: error message, 500: Internal Server Error - if unexpected error occurred during query
+            (indicator that something is wrong with server code)
         """
         try:
             res = db_helper.get_statistic_for_import_id(import_id)
@@ -210,5 +209,3 @@ def create_app(test_config=None):
             return return_str, 500
 
     return app
-
-
